fix the connection event from socket.io not firing sometimes.
